Moves cracked key methods to own class.
